# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/billing/plans": {
      "post": {
        "operationId": "BillingController_createPlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 201,
                    "message": "Plan created successfully",
                    "data": {
                      "id": "plan-id",
                      "name": "Pro",
                      "price": 29.99,
                      "currency": "USD",
                      "interval": "MONTHLY"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a subscription plan (Super Admin only)",
        "tags": [
          "Billing"
        ]
      },
      "get": {
        "operationId": "BillingController_getPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Plans retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Plans retrieved successfully",
                    "data": [
                      {
                        "id": "plan-id",
                        "name": "Pro",
                        "price": 29.99,
                        "currency": "USD",
                        "interval": "MONTHLY"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List all active subscription plans",
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/plans/{id}": {
      "patch": {
        "operationId": "BillingController_updatePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Plan updated successfully",
                    "data": {
                      "id": "plan-id",
                      "name": "Pro Updated",
                      "price": 39.99
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update a subscription plan (Super Admin only)",
        "tags": [
          "Billing"
        ]
      }
    },
    "/tenants/{tenantId}/billing/subscribe": {
      "post": {
        "operationId": "BillingController_subscribe",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscribed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Subscribed successfully",
                    "data": {
                      "id": "sub-id",
                      "tenantId": "tenant-id",
                      "planId": "plan-id",
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Subscribe to a plan (Owner only)",
        "tags": [
          "Billing"
        ]
      }
    },
    "/tenants/{tenantId}/billing/subscription": {
      "get": {
        "operationId": "BillingController_getSubscription",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription details returned",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Subscription retrieved successfully",
                    "data": {
                      "id": "sub-id",
                      "planId": "plan-id",
                      "status": "ACTIVE",
                      "currentPeriodEnd": "2026-07-16T00:00:00Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get active subscription details",
        "tags": [
          "Billing"
        ]
      }
    },
    "/tenants/{tenantId}/billing/cancel": {
      "post": {
        "operationId": "BillingController_cancelSubscription",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Subscription cancelled successfully",
                    "data": {
                      "id": "sub-id",
                      "status": "CANCELED"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Cancel active subscription (Owner only)",
        "tags": [
          "Billing"
        ]
      }
    },
    "/tenants/{tenantId}/billing/invoices": {
      "get": {
        "operationId": "BillingController_getInvoices",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoices retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Invoices retrieved successfully",
                    "data": [
                      {
                        "id": "invoice-id",
                        "amount": 29.99,
                        "status": "PAID",
                        "issuedAt": "2026-06-16T00:00:00Z"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List all invoices for the workspace (Owner only)",
        "tags": [
          "Billing"
        ]
      }
    },
    "/tenants/{tenantId}/billing/invoices/{id}": {
      "get": {
        "operationId": "BillingController_getInvoiceById",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details returned",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Invoice retrieved successfully",
                    "data": {
                      "id": "invoice-id",
                      "amount": 29.99,
                      "status": "PAID",
                      "issuedAt": "2026-06-16T00:00:00Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get details of a specific invoice (Owner only)",
        "tags": [
          "Billing"
        ]
      }
    }
  },
  "info": {
    "title": "SaaS Multi-Tenant API",
    "description": "Auto-generated OpenAPI specification for the SaaS Demo application.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token to authenticate",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "billingCycle": {
            "type": "string",
            "description": "e.g., MONTHLY or YEARLY"
          },
          "features": {
            "type": "object"
          },
          "stripeProductId": {
            "type": "string"
          },
          "stripePriceId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "price",
          "billingCycle"
        ]
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string"
          }
        },
        "required": [
          "planId"
        ]
      }
    }
  }
}